Replace long long with int64_t. Use separate reference to db and separate SQLite statements
Updates include:
 * SearchMarkerQuery: use separate reference to db and separate SQLite
   statements
 * Replace long long casts with int64_t
 * PositionQuery.cpp: fix potential overflow
 * Repository.cpp:
     - Fix possible race condition when opening the database
     - Handle bbox crossing antimeridian
 * SqliteCppUtil: downgrade assert to warning

+++ Queries/AddressQuery.cpp
@@ -75,7 +75,7 @@ bool AddressQuery::Delete(const ACDB_marker_idx_type aId) {
   bool success = false;
 
   try {
-    mDelete->bind(Parameters::Id, static_cast<long long>(aId));
+    mDelete->bind(Parameters::Id, static_cast<int64_t>(aId));
 
     mDelete->exec();
     success = mDelete->isDone();
@@ -105,8 +105,8 @@ bool AddressQuery::Delete(const uint64_t aGeohashStart, const uint64_t aGeohashE
   bool success = false;
 
   try {
-    mDeleteGeohash->bind(Parameters::GeohashStart, static_cast<long long>(aGeohashStart));
-    mDeleteGeohash->bind(Parameters::GeohashEnd, static_cast<long long>(aGeohashEnd));
+    mDeleteGeohash->bind(Parameters::GeohashStart, static_cast<int64_t>(aGeohashStart));
+    mDeleteGeohash->bind(Parameters::GeohashEnd, static_cast<int64_t>(aGeohashEnd));
 
     mDeleteGeohash->exec();
     success = mDeleteGeohash->isDone();
@@ -137,7 +137,7 @@ bool AddressQuery::Get(const ACDB_marker_idx_type aId, AddressTableDataType& aRe
   bool success = false;
 
   try {
-    mRead->bind(Parameters::Id, static_cast<long long>(aId));
+    mRead->bind(Parameters::Id, static_cast<int64_t>(aId));
 
     success = mRead->executeStep();
     if (success) {
@@ -171,7 +171,7 @@ bool AddressQuery::Write(const ACDB_marker_idx_type aId, AddressTableDataType&& 
   bool success = false;
 
   try {
-    mWrite->bind(Parameters::Id, static_cast<long long>(aId));
+    mWrite->bind(Parameters::Id, static_cast<int64_t>(aId));
     mWrite->bind(Parameters::SectionTitle, aAddressTableData.mSectionTitle);
     mWrite->bind(Parameters::String, aAddressTableData.mStringFieldsJson);
     mWrite->bind(Parameters::Labeled, aAddressTableData.mAttributeFieldsJson);
